Use OpenGL 2.1 on desktop, OpenGL ES 2.0 on mobile.
Rewrite the build constraints so that the OpenGL 2.1 backend based on
go-gl/gl/v2.1/gl package is used on desktop (OS X, Linux and Windows).

Use the original x/mobile/gl OpenGL ES 2.0 backend only on mobile
platforms (iOS and Android).
